requirements.txt
Django==1.9.7
django-bootstrap3==7.0.1
Pillow==3.3.0